Refactor table matrix
