+++ src/components/SlateEditor/plugins/table/helpers.ts
@@ -113,3 +113,17 @@ export const updateCell = (
 export const getPrevCell = (matrix: TableMatrix, row: number, column: number) => {
   return matrix[row][column - 1];
 };
+
+// Find the matrix coordinates for a cell. Returns the coordinates for top left corner of cell.
+export const findCellCoordinate = (
+  matrix: TableMatrix,
+  targetCell: TableCellElement,
+): [number, number] | undefined => {
+  for (const [rowIndex, row] of matrix.entries()) {
+    for (const [cellIndex, cell] of row.entries()) {
+      if (cell === targetCell) {
+        return [rowIndex, cellIndex];
+      }
+    }
+  }
+};
